RythmGame
This project is a rythm game that I created with Unity and C# as part of the "Nuit de l'Info". This is a French event where students must develop a website around a theme and complete challenges. One challenge was to develop a rythm game and integrate it to the website. So I volunteered to make this game during the night.

It can definitely be improved but I didn't really have time for details so I might update it later.